Separated overlays into several files
